requirements.txt
coverage==3.7.1
mock==1.0.1
nose==1.3.1
PyYAML==3.11
juju-deployer==0.3.5
